tests: refactor jvm tests to not depend on other tests
