+++ src/python/pants/jvm/shading/rules_integration_test.py
@@ -3,7 +3,6 @@
 
 from __future__ import annotations
 
-import dataclasses
 import os
 
 import pytest
@@ -16,7 +15,7 @@ from pants.backend.java.dependency_inference.rules import rules as java_dep_inf_
 from pants.backend.java.target_types import rules as target_types_rules
 from pants.core.util_rules import archive
 from pants.core.util_rules.archive import ExtractedArchive, MaybeExtractArchiveRequest
-from pants.engine.fs import EMPTY_DIGEST, AddPrefix, CreateDigest, Digest, DigestContents, Snapshot
+from pants.engine.fs import EMPTY_DIGEST, AddPrefix, Digest, Snapshot
 from pants.engine.target import AllTargets, CoarsenedTargets, CoarsenedTargetsRequest
 from pants.jvm import classpath
 from pants.jvm import compile as jvm_compile
@@ -35,7 +34,7 @@ from pants.jvm.target_types import (
     JvmShadingRelocateRule,
     JvmShadingRenameRule,
 )
-from pants.jvm.testutil import maybe_skip_jdk_test
+from pants.jvm.testutil import _get_jar_contents_snapshot, maybe_skip_jdk_test
 from pants.jvm.util_rules import rules as jvm_util_rules
 from pants.testutil.rule_runner import PYTHON_BOOTSTRAP_ENV, QueryRule, RuleRunner
 
@@ -113,22 +112,6 @@ def _resolve_jar(rule_runner: RuleRunner, coord: Coordinate) -> ClasspathEntry:
     return rule_runner.request(
         ClasspathEntry, [CoursierFetchRequest(coarsened_tgts[0], resolve=resolve_key)]
     )
-
-
-def _get_jar_contents_snapshot(
-    rule_runner: RuleRunner, *, filename: str, digest: Digest
-) -> Snapshot:
-    contents = rule_runner.request(DigestContents, [digest])
-    files_content = [content for content in contents if content.path == filename]
-    assert len(files_content) == 1
-
-    renamed_digest = rule_runner.request(
-        Digest, [CreateDigest([dataclasses.replace(files_content[0], path=f"{filename}.zip")])]
-    )
-    extracted_jar = rule_runner.request(
-        ExtractedArchive, [MaybeExtractArchiveRequest(renamed_digest)]
-    )
-    return rule_runner.request(Snapshot, [extracted_jar.digest])
 
 
 @maybe_skip_jdk_test
